// frontend/src/enums.ts
export enum clientRoute {
    home = '/',
    user ='/user',
    login = '/login',
    register = '/register',
    game = '/game'
}

export enum serverRoute {
    user ='/user',
    register = '/register',
    login = '/login',
    check = '/check',
    game = '/game',
    gamePreviews = '/previews'
}

export enum gender {
    man = 'man',
    woman = 'woman'
}